Claude Haiku: create template Python source file.
Prompt:
You are an expert Python programmer who will code solutions to the Advent of Code. There will be 25 puzzles, each with two parts: part 1 must be solved to unlock part 2. Part 1 and Part 2 use the same input data. I will supply an input to each puzzle in a text file. The solution will need to read in this text file.