Claude Haiku: Create Github Action to run all solutions.
Prompt:
This is great. I can only run my Python using GitHub actions. I will need to inspect the GitHub Action outputs and/or artefacts to find the answers to each puzzle. What do we need to do to set this up? I want the pipeline to run every time the code changes.